# dist/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Coppia.co API Documentation",
        "description": "API Documentation for Coppia.co API calls",
        "termsOfService": "http://swagger.io/terms/",
        "contact" : {
            "name" : "Coppia API Support",
            "url" : "www.coppia.co/support",
            "email" : "eric@coppia.co"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
...
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/v1/customers/": {
            "get": {
                "tags": ["Customers"],
                "description": "Return list of Customers",
                "operationId": "GetCustomers",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Customer object returned in list",
                        "schema": {
                            "$ref": "#/definitions/customer"
                        }
                    }
                }
            },
            "post":{
                "tags": [
                    "Customers"
                ],
                "description": "Creates a new Customer",
                "operationId": "CreateCustomer",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "body",
                        "description": "First Name of the Customer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "body",
                        "description": "Last Name of the Customer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "body",
                        "description": "Email Address of the Customer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "image_link",
                        "in": "body",
                        "description": "Image/Avatar url of the Customer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "body",
                        "description": "Title/Position of the Customer",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "company_name",
                        "in": "body",
                        "description": "Company of the Customer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "create_user",
                        "in": "body",
                        "description": "User Id for the logged in user",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "When creating a new customer, the id for the newly created Customer will be returned",
                        "schema": {
                            "$ref": "#/definitions/create_customer"
                        }
                    }
                }
            }
        },
        "/api/v1/customers/{customer_id}": {
            "get":{
                "tags": [
                    "Customers"
                ],
                "description": "Returns a single Customer object",
                "operationId": "GetCustomerById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "description": "id of the Customer to return",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns single Customer object",
                        "schema": {
                            "$ref": "#/definitions/customer"
                        }
                    }
                }
            },
            "put":{
                "tags": [
                    "Customers"
                ],
                "description": "Returns a success message with count of customer updates",
                "operationId": "UpdateCustomerById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "description": "id for the customer to update",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    },
                    {
                        "name": "first_name",
                        "in": "body",
                        "description": "First Name of the Customer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "body",
                        "description": "Last Name of the Customer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "body",
                        "description": "Email Address of the Customer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "image_link",
                        "in": "body",
                        "description": "Image/Avatar url of the Customer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "update_user",
                        "in": "body",
                        "description": "id for the logged in user",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response when calling to update a Customer",
                        "schema": {
                            "$ref": "#/definitions/update_customer"
                        }
                    }
                }
            },
            "delete":{
                "tags":[
                    "Customers"
                ],
                "description":"Deletes a customer from the database",
                "operationId":"DeleteCustomer",
                "produces": [
                    "application/json"
                ],
                "parameters":[
                    {
                        "name": "customer_id",
                        "in": "path",
                        "description": "customer_id of the Customer to be deleted from the database.",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response for the deleted Customer",
                        "schema":{
                            "$ref":"#/definitions/delete_customer"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "customer": {
            "required": [
                "id",
                "first_name",
                "last_name",
                "email"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format" : "int32",
                    "example": "6"
                },
                "first_name": {
                    "type": "string",
                    "example": "Eric"
                },
                "last_name": {
                    "type": "string",
                    "example": "Hillenbrand"
                },
                "email": {
                    "type": "string",
                    "example": "eric@coppia.co"
                },
                "image_link": {
                    "type": "string",
                    "example": "https://d1ts43dypk8bqh.cloudfront.net/v1/avatars/88a1184e-964b-495b-86fc-92c2867c5427"
                },
                "title": {
                    "type": "string",
                    "example": "Product Marketing Manager"
                },
                "company_name": {
                    "type": "string",
                    "example": "Web PT"
                },
                "created_by": {
                    "type": "string",
                    "example": "Eric Hillenbrand"
                },
                "created_date": {
                    "type": "string",
                    "format" : "date-time",
                    "example": "2016-10-17T23:10:23.000Z"
                },
                "created_image_link": {
                    "type": "string",
                    "example": "https://d1ts43dypk8bqh.cloudfront.net/v1/avatars/88a1184e-964b-495b-86fc-92c2867c5427"
                },
                "updated_by": {
                    "type": "string",
                    "example": "Eric Hillenbrand"
                },
                "updated_datetime": {
                    "type": "string",
                    "format" : "date-time",
                    "example": "2016-10-17T23:10:23.000Z"
                },
                "updated_image_link": {
                    "type": "string",
                    "example": "https://d1ts43dypk8bqh.cloudfront.net/v1/avatars/88a1184e-964b-495b-86fc-92c2867c5427"
                }
            }
        },
        "create_customer": {
            "properties": {
                "customer_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example" : "3"
                }
            }
        },
         "update_customer": {
            "properties": {
                "customer_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example" : "8"
                }
            }
        },
         "delete_customer": {
            "properties": {
                "customer_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example" : "6"
                }
            }
        },
...
    }
}